third_party/blink/web_tests/svg/parser/whitespace-angle-1-expected.txt

CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-47a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".1a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "0.35a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "1e-10a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-47e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".1e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "0.35e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "1e-10e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-47foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".1foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "0.35foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "1e-10foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-47)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".1)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "0.35)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "1e-10)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-47 deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".1 deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "0.35 deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "1e-10 deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-47 rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".1 rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "0.35 rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "1e-10 rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-47 grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".1 grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "0.35 grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "1e-10 grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -47a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .1a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " 0.35a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " 1e-10a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -47e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .1e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " 0.35e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " 1e-10e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -47foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .1foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " 0.35foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " 1e-10foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -47)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .1)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " 0.35)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " 1e-10)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-47   deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".1   deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "0.35   deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "1e-10   deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-47   rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".1   rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "0.35   rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "1e-10   rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-47   grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".1   grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "0.35   grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "1e-10   grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -47a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .1a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   0.35a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   1e-10a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -47e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .1e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   0.35e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   1e-10e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -47foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .1foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   0.35foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   1e-10foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -47)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .1)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   0.35)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   1e-10)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-47\r\n\t deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".1\r\n\t deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "0.35\r\n\t deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "1e-10\r\n\t deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-47\r\n\t rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".1\r\n\t rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "0.35\r\n\t rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "1e-10\r\n\t rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-47\r\n\t grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".1\r\n\t grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "0.35\r\n\t grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "1e-10\r\n\t grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -47a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .1a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t 0.35a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t 1e-10a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -47e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .1e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t 0.35e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t 1e-10e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -47foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .1foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t 0.35foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t 1e-10foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -47)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .1)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t 0.35)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t 1e-10)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-47\fdeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".1\fdeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "0.35\fdeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "1e-10\fdeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-47\frad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".1\frad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "0.35\frad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "1e-10\frad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-47\fgrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".1\fgrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "0.35\fgrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "1e-10\fgrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-47a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.1a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f0.35a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f1e-10a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-47e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.1e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f0.35e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f1e-10e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-47foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.1foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f0.35foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f1e-10foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-47)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.1)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f0.35)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f1e-10)90".
This is a testharness.js-based test.
Harness: the test ran to completion.
